+++ portal/models/user_consent.py
@@ -5,7 +5,7 @@ from validators import url as url_validation
 
 from .audit import Audit
 from ..database import db
-from .fhir import FHIR_datetime
+from ..date_tools import FHIR_datetime
 from .organization import Organization
 from .user import User
